* 'teature/member' of https://github.com/me-coldrain/BE:
  Fix: "고침 팀원 참여 승인/거절 API" 팀 개설자만 승인/거절을 할 수 있도록 변경
  Fix: "고침 팀 등록 API" S3에 이미지 업로드 로직 추가, 회원에 등록한 팀 설정

+++ src/main/java/me/coldrain/ninetyminute/service/ParticipationService.java
@@ -2,8 +2,10 @@ package me.coldrain.ninetyminute.service;
 
 import lombok.RequiredArgsConstructor;
 import me.coldrain.ninetyminute.dto.request.TeamParticipateRequest;
+import me.coldrain.ninetyminute.entity.Member;
 import me.coldrain.ninetyminute.entity.Participation;
 import me.coldrain.ninetyminute.entity.Team;
+import me.coldrain.ninetyminute.repository.MemberRepository;
 import me.coldrain.ninetyminute.repository.ParticipationRepository;
 import me.coldrain.ninetyminute.repository.TeamRepository;
 import org.springframework.stereotype.Service;
@@ -16,15 +18,19 @@ public class ParticipationService {
 
     private final ParticipationRepository participationRepository;
     private final TeamRepository teamRepository;
+    private final MemberRepository memberRepository;
 
     @Transactional
-    public void participate(final Long teamId, final TeamParticipateRequest request) {
+    public void participate(
+            final Long teamId,
+            final Member member,
+            final TeamParticipateRequest request) {
+
         final Team team = teamRepository.findById(teamId)
                 .orElseThrow(() -> new IllegalArgumentException("참여하려는 팀이 존재하지 않습니다."));
 
-        // TODO: 2022-07-07 로그인한 사용자를 주입 하도록 코드 변경 필요
         final Participation participation = Participation.builder()
-                .member(null)
+                .member(member)
                 .team(team)
                 .approved(false)
                 .answer(request.getAnswer())
@@ -34,17 +40,27 @@ public class ParticipationService {
     }
 
     @Transactional
-    public void approve(final Long teamId, final Long memberId) {
+    public void approve(final Long teamId, final Long memberId, final Member member) {
         final Participation participation = participationRepository.findByTeamIdAndMemberId(teamId, memberId)
                 .orElseThrow(() -> new IllegalArgumentException("참여 이력이 존재하지 않습니다."));
+
+        final Long openTeamId = member.getOpenTeam().getId();
+        if (!openTeamId.equals(teamId)) {
+            throw new IllegalArgumentException("팀 개설자만 참여 승인을 할 수 있습니다.");
+        }
 
         participation.changeApproved(true);
     }
 
     @Transactional
-    public void disapprove(final Long teamId, final Long memberId) {
+    public void disapprove(final Long teamId, final Long memberId, final Member member) {
         final Participation participation = participationRepository.findByTeamIdAndMemberId(teamId, memberId)
                 .orElseThrow(() -> new IllegalArgumentException("참여 이력이 존재하지 않습니다."));
+
+        final Long openTeamId = member.getOpenTeam().getId();
+        if (!openTeamId.equals(teamId)) {
+            throw new IllegalArgumentException("팀 개설자만 참여 승인을 할 수 있습니다.");
+        }
 
         participationRepository.delete(participation);
     }
